Implement DSNCommand for read- and delivery receipt
We use the Delivery Status Notification (independently on whether
they are configured for the server or smtp client) to ask for a
delivery or read receipt.

This can be done by adding `Disposition-Notification-To` or
`Read-Receipt-To` to the header fields in combination with a
return e-mail address that receives the notification.

This command should make it easily configurable so that the user
can simply call: `receipt-read` or `receipt-delivered` in the
envelope mode and a field is correctly added to the header.